external/DJtransGAN/requirements_patched.txt
#absl-py==0.13.0
acoustics==0.2.6
asteroid-filterbanks==0.4.0
asttokens==3.0.1
audioread==3.0.1
backcall==0.2.0
beautifulsoup4==4.14.3
# certifi==2025.11.12
# cffi==1.17.1
# charset-normalizer==3.4.4
colorama==0.4.6
cycler==0.12.1
# Cython==0.29.36
# decorator==5.2.1
executing==2.2.1
# filelock==3.16.1
# future==1.0.0
gdown==5.2.0
# idna==3.11
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipython==8.12.3
jedi==0.19.2
# joblib==1.4.2
# kiwisolver==1.4.7
# librosa==0.8.1
# llvmlite==0.41.1
# madmom @ git+https://github.com/CPJKU/madmom.git
# matplotlib==3.4.2
matplotlib-inline==0.1.7
# mido==1.3.3
nnAudio==0.3.3
# numba==0.58.1
# numpy==1.22.4
openunmix==1.2.1
# packaging==25.0
# pandas==2.0.3
parso==0.8.5
pickleshare==0.7.5
# pillow==10.4.0
# platformdirs==4.3.6
# pooch==1.8.2
prompt_toolkit==3.0.52
pure_eval==0.2.3
# pycparser==2.23
# Pygments==2.19.2
pyloudnorm==0.1.1
# pyparsing==3.1.4
pyrubberband==0.4.0
PySocks==1.7.1
# python-dateutil==2.9.0.post0
# pytz==2025.2
# PyYAML==5.4.1
# requests==2.32.4
resampy==0.4.3
# scikit-learn==1.3.2
# scipy==1.7.3
# six==1.17.0
# soundfile==0.13.1
soupsieve==2.7
stack-data==0.6.3
tabulate==0.9.0
# threadpoolctl==3.5.0
# torch==1.10.2
# torchaudio==0.10.2
torchlibrosa==0.1.0
# torchvision==0.11.3
# tqdm==4.67.1
traitlets==5.14.3
# typing_extensions==4.13.2
# tzdata==2025.2
# urllib3==2.2.3
wcwidth==0.2.14
zipp==3.20.2


# Run for Madmom Installation: python -m pip install --no-build-isolation "madmom @ git+https://github.com/CPJKU/madmom.git@v0.16.1"



